Add GA4 event tracking and improve per-page SEO
GA4 event tracking:
- Create gtag utility (src/lib/gtag.js) with helpers for all key events
- Track contact form submissions as generate_lead events
- Track phone, WhatsApp, and email clicks on contact page
- Track brochure PDF downloads as file_download events
- Track CTA button clicks on homepage and services page
- Track phone click on about page
- Track social media link clicks in footer

SEO improvements (non-breaking):
- Add robots.js for proper robots.txt generation with sitemap reference
- Add brochure page to sitemap.js
- Restructure all pages to server components with per-page metadata
  (title, description, canonical URL) by extracting client code into
  separate component files
- Remove duplicate hardcoded canonical link tag from layout (handled
  by Next.js metadata API)
