+++ src/app/gallery/page.js
@@ -1,226 +1,17 @@
-'use client';
+import GalleryContent from "@/component/GalleryContent";
 
-import { motion, AnimatePresence } from 'framer-motion';
-import Image from 'next/image';
-import { useState } from 'react';
-import { X, ChevronLeft, ChevronRight } from 'lucide-react';
+export const metadata = {
+  title: "Gallery | Pimp My Party - Event Photos & Highlights Manchester",
+  description: "Browse photos from weddings, parties and events across Manchester and Greater Manchester. See our DJ setups, saxophone performances, photobooths and venue lighting in action.",
+  alternates: {
+    canonical: '/gallery',
+  },
+};
 
 export default function GalleryPage() {
-  const [selectedImage, setSelectedImage] = useState(null);
-  const [currentIndex, setCurrentIndex] = useState(0);
-
-  // All your gallery images
-  const images = [
-    '/gallery/bride_groom_love_decor.png',
-    '/gallery/bride_groom.jpeg',
-    '/gallery/cold_spark_machine_2.jpg',
-    '/gallery/cold_spark_machine.webp',
-    '/gallery/first_dance.jpeg',
-    '/gallery/original.webp',
-    '/gallery/sax_player_2.jpg',
-    '/gallery/sax_player_3.jpg',
-    '/gallery/sax_player_and_bride.jpeg',
-    '/gallery/sax_player.jpg',
-    '/gallery/wedding_dance.webp',
-    '/gallery/wedding_party.jpg',
-    '/gallery/wedding_venue_2.png',
-    '/gallery/wedding_venue.jpg'
-  ];
-
-  const openLightbox = (index) => {
-    setCurrentIndex(index);
-    setSelectedImage(images[index]);
-  };
-
-  const closeLightbox = () => {
-    setSelectedImage(null);
-  };
-
-  const goToNext = () => {
-    const nextIndex = (currentIndex + 1) % images.length;
-    setCurrentIndex(nextIndex);
-    setSelectedImage(images[nextIndex]);
-  };
-
-  const goToPrevious = () => {
-    const prevIndex = (currentIndex - 1 + images.length) % images.length;
-    setCurrentIndex(prevIndex);
-    setSelectedImage(images[prevIndex]);
-  };
-
   return (
-    <div className="min-h-screen bg-gradient-to-b from-gray-900 via-gray-900 to-black text-white">
-      
-      {/* Header */}
-      <section className="py-12 px-4">
-        <motion.div
-          initial={{ opacity: 0, y: 30 }}
-          animate={{ opacity: 1, y: 0 }}
-          transition={{ duration: 0.8 }}
-          className="max-w-7xl mx-auto text-center"
-        >
-          <h1 className="text-5xl md:text-6xl font-bold mb-4 pb-2 bg-gradient-to-r from-purple-400 via-pink-400 to-purple-400 bg-clip-text text-transparent">
-            Our Gallery
-          </h1>
-          <motion.div 
-            className="w-32 h-1 bg-gradient-to-r from-purple-500 via-pink-500 to-purple-500 mx-auto mb-6"
-            initial={{ scaleX: 0 }}
-            animate={{ scaleX: 1 }}
-            transition={{ duration: 0.8, delay: 0.3 }}
-          />
-          <p className="text-xl text-gray-400">Moments we've created</p>
-        </motion.div>
-      </section>
-
-      {/* Gallery Grid */}
-      <section className="py-8 px-4">
-        <div className="max-w-7xl mx-auto">
-          <motion.div 
-            className="grid grid-cols-2 md:grid-cols-3 lg:grid-cols-4 gap-4"
-            initial="hidden"
-            animate="visible"
-            variants={{
-              visible: {
-                transition: {
-                  staggerChildren: 0.1
-                }
-              }
-            }}
-          >
-            {images.map((image, index) => (
-              <motion.div
-                key={index}
-                variants={{
-                  hidden: { opacity: 0, scale: 0.8 },
-                  visible: { opacity: 1, scale: 1 },
-                  transition: { duration: 0.6 }
-                }}
-                whileHover={{ 
-                  scale: 1.05,
-                  zIndex: 10,
-                  transition: { duration: 0.5 }
-                }}
-                className="relative aspect-square cursor-pointer overflow-hidden rounded-xl group"
-                onClick={() => openLightbox(index)}
-              >
-                {/* Image */}
-                <Image
-                  src={image}
-                  alt={`Gallery image ${index + 1}`}
-                  fill
-                  className="object-cover"
-                  sizes="(max-width: 768px) 50vw, (max-width: 1024px) 33vw, 25vw"
-                />
-                
-                {/* Gradient overlay on hover */}
-                <motion.div
-                  className="absolute inset-0 bg-gradient-to-t from-purple-900/80 via-transparent to-transparent opacity-0 group-hover:opacity-100 transition-opacity duration-300"
-                />
-                
-                {/* Glimmer effect on hover */}
-                <motion.div
-                  className="absolute inset-0 bg-gradient-to-r from-transparent via-white/20 to-transparent"
-                  initial={{ x: '-100%' }}
-                  whileHover={{ x: '100%' }}
-                  transition={{ duration: 0.6 }}
-                />
-
-                {/* Border glow on hover */}
-                <div className="absolute inset-0 border-2 border-transparent group-hover:border-purple-400 rounded-xl transition-all duration-300" />
-              </motion.div>
-            ))}
-          </motion.div>
-        </div>
-      </section>
-
-      {/* Lightbox Modal */}
-      <AnimatePresence>
-        {selectedImage && (
-          <motion.div
-            initial={{ opacity: 0 }}
-            animate={{ opacity: 1 }}
-            exit={{ opacity: 0 }}
-            className="fixed inset-0 bg-black/95 z-50 flex items-center justify-center p-4"
-            onClick={closeLightbox}
-          >
-            {/* Close button */}
-            <motion.button
-              initial={{ opacity: 0, scale: 0.8 }}
-              animate={{ opacity: 1, scale: 1 }}
-              exit={{ opacity: 0, scale: 0.8 }}
-              whileHover={{ scale: 1.1, rotate: 90 }}
-              whileTap={{ scale: 0.9 }}
-              className="absolute top-4 right-4 z-50 bg-purple-600 hover:bg-purple-700 text-white rounded-full p-2 transition-colors"
-              onClick={closeLightbox}
-            >
-              <X size={32} />
-            </motion.button>
-
-            {/* Previous button */}
-            <motion.button
-              initial={{ opacity: 0, x: -20 }}
-              animate={{ opacity: 1, x: 0 }}
-              whileHover={{ scale: 1.1, x: -5 }}
-              whileTap={{ scale: 0.9 }}
-              className="absolute left-4 z-50 bg-purple-600 hover:bg-purple-700 text-white rounded-full p-3 transition-all"
-              onClick={(e) => {
-                e.stopPropagation();
-                goToPrevious();
-              }}
-            >
-              <ChevronLeft size={32} />
-            </motion.button>
-
-            {/* Next button */}
-            <motion.button
-              initial={{ opacity: 0, x: 20 }}
-              animate={{ opacity: 1, x: 0 }}
-              whileHover={{ scale: 1.1, x: 5 }}
-              whileTap={{ scale: 0.9 }}
-              className="absolute right-4 z-50 bg-purple-600 hover:bg-purple-700 text-white rounded-full p-3 transition-all"
-              onClick={(e) => {
-                e.stopPropagation();
-                goToNext();
-              }}
-            >
-              <ChevronRight size={32} />
-            </motion.button>
-
-            {/* Image container */}
-            <motion.div
-              key={selectedImage}
-              initial={{ scale: 0.8, opacity: 0, rotateY: -15 }}
-              animate={{ scale: 1, opacity: 1, rotateY: 0 }}
-              exit={{ scale: 0.8, opacity: 0, rotateY: 15 }}
-              transition={{ duration: 0.6 }}
-              className="relative max-w-5xl max-h-[90vh] w-full h-full"
-              onClick={(e) => e.stopPropagation()}
-            >
-              {/* Glowing border effect */}
-              <div className="absolute -inset-1 bg-gradient-to-r from-purple-600 via-pink-600 to-purple-600 rounded-lg blur-lg opacity-75 animate-pulse" />
-              
-              <div className="relative w-full h-full rounded-lg overflow-hidden">
-                <Image
-                  src={selectedImage}
-                  alt="Selected gallery image"
-                  fill
-                  className="object-contain"
-                  sizes="90vw"
-                />
-              </div>
-            </motion.div>
-
-            {/* Image counter */}
-            <motion.div
-              initial={{ opacity: 0, y: 20 }}
-              animate={{ opacity: 1, y: 0 }}
-              className="absolute bottom-8 left-1/2 transform -translate-x-1/2 bg-purple-600/80 backdrop-blur-sm px-6 py-3 rounded-full text-white font-semibold"
-            >
-              {currentIndex + 1} / {images.length}
-            </motion.div>
-          </motion.div>
-        )}
-      </AnimatePresence>
-    </div>
+    <main>
+      <GalleryContent />
+    </main>
   );
 }
