+++ src/app/robots.js
@@ -1,0 +1,12 @@
+export default function robots() {
+  return {
+    rules: [
+      {
+        userAgent: '*',
+        allow: '/',
+        disallow: ['/api/', '/_next/'],
+      },
+    ],
+    sitemap: 'https://pimpmyparty.co.uk/sitemap.xml',
+  };
+}
